Remove the switchType junk
