+++ api/src/main/java/com/dsh105/echopet/compat/api/entity/GuardianType.java
@@ -1,4 +1,4 @@
-package net.techcable.sonarpet.nms.switching;
+package com.dsh105.echopet.compat.api.entity;
 
 public enum GuardianType {
     ELDER,
