+++ api/src/main/java/com/dsh105/echopet/compat/api/entity/type/pet/IZombiePet.java
@@ -19,7 +19,7 @@ package com.dsh105.echopet.compat.api.entity.type.pet;
 
 import com.dsh105.echopet.compat.api.entity.IAgeablePet;
 
-import net.techcable.sonarpet.nms.switching.ZombieType;
+import com.dsh105.echopet.compat.api.entity.ZombieType;
 
 public interface IZombiePet extends IAgeablePet {
