+++ core/src/main/java/net/techcable/sonarpet/nms/entity/type/EntityCaveSpiderPet.java
@@ -11,20 +11,12 @@ import net.techcable.sonarpet.nms.NMSInsentientEntity;
 
 @EntityPetType(petType = PetType.CAVESPIDER)
 public class EntityCaveSpiderPet extends EntityInsentientPet implements IEntityCaveSpiderPet {
-    private final NMSInsentientEntity entity;
-
     protected EntityCaveSpiderPet(IPet pet, NMSInsentientEntity entity) {
-        super(pet);
-        this.entity = entity;
+        super(pet, entity);
     }
 
     @Override
     public void makeStepSound() {
         getEntity().playSound(SafeSound.SPIDER_STEP, 0.15F, 1.0F);
-    }
-
-    @Override
-    public NMSInsentientEntity getEntity() {
-        return entity;
     }
 }
