+++ core/src/main/java/net/techcable/sonarpet/nms/entity/type/EntityChickenPet.java
@@ -11,20 +11,12 @@ import net.techcable.sonarpet.nms.NMSInsentientEntity;
 
 @EntityPetType(petType = PetType.CHICKEN)
 public class EntityChickenPet extends EntityAgeablePet implements IEntityChickenPet {
-    private final NMSInsentientEntity entity;
-
     protected EntityChickenPet(IPet pet, NMSInsentientEntity entity) {
-        super(pet);
-        this.entity = entity;
+        super(pet, entity);
     }
 
     @Override
     public void makeStepSound() {
         getEntity().playSound(SafeSound.CHICKEN_STEP, 0.15F, 1.0F);
-    }
-
-    @Override
-    public NMSInsentientEntity getEntity() {
-        return entity;
     }
 }
