+++ core/src/main/java/net/techcable/sonarpet/nms/entity/type/EntityEndermanPet.java
@@ -13,21 +13,14 @@ import org.bukkit.entity.Enderman;
 
 @EntityPetType(petType = PetType.ENDERMAN)
 public class EntityEndermanPet extends EntityNoClipPet implements IEntityEndermanPet {
-    private final NMSInsentientEntity entity;
     protected EntityEndermanPet(IPet pet, NMSInsentientEntity entity) {
-        super(pet);
-        this.entity = entity;
+        super(pet, entity);
     }
     private static final int SCREAMING_METADATA_ID = Versioning.NMS_VERSION.getMetadataId("ENDERMAN_SCREAMING_METADATA_ID");
 
     @Override
     public void setScreaming(boolean flag) {
         getEntity().getDataWatcher().setBoolean(SCREAMING_METADATA_ID, flag);
-    }
-
-    @Override
-    public NMSInsentientEntity getEntity() {
-        return entity;
     }
 
     @Override
