+++ core/src/main/java/net/techcable/sonarpet/nms/entity/type/EntityGuardianPet.java
@@ -5,12 +5,9 @@ import com.dsh105.echopet.compat.api.entity.IPet;
 import com.dsh105.echopet.compat.api.entity.PetType;
 import com.dsh105.echopet.compat.api.entity.SizeCategory;
 import com.dsh105.echopet.compat.api.entity.type.nms.IEntityGuardianPet;
-import net.techcable.sonarpet.nms.INMS;
 
 import net.techcable.sonarpet.nms.entity.EntityInsentientPet;
 import net.techcable.sonarpet.nms.NMSInsentientEntity;
-import net.techcable.sonarpet.nms.switching.EntitySwitchReason;
-import net.techcable.sonarpet.nms.switching.GuardianType;
 import net.techcable.sonarpet.utils.NmsVersion;
 import net.techcable.sonarpet.utils.Versioning;
 
@@ -19,21 +16,14 @@ import org.bukkit.entity.Guardian;
 
 @EntityPetType(petType = PetType.GUARDIAN)
 public class EntityGuardianPet extends EntityInsentientPet implements IEntityGuardianPet {
-    private NMSInsentientEntity entity;
     protected EntityGuardianPet(IPet pet, NMSInsentientEntity entity) {
-        super(pet);
-        this.entity = entity;
+        super(pet, entity);
     }
 
     @Override
     public SizeCategory getSizeCategory() {
         return isElder() ? SizeCategory.GIANT : SizeCategory.LARGE;
 
-    }
-
-    @Override
-    public NMSInsentientEntity getEntity() {
-        return entity;
     }
 
     @Override
@@ -50,11 +40,7 @@ public class EntityGuardianPet extends EntityInsentientPet implements IEntityGua
     @SuppressWarnings("deprecation") // Only done on versions it's safe for
     public void setElder(boolean flag) {
         if (Versioning.NMS_VERSION.compareTo(NmsVersion.v1_11_R1) >= 0) {
-            entity = INMS.getInstance().switchType(
-                    entity,
-                    EntitySwitchReason.GUARDIAN_SWITCH,
-                    flag ? GuardianType.ELDER : GuardianType.REGULAR
-            );
+            throw new UnsupportedOperationException("TODO");
         } else {
             getBukkitEntity().setElder(flag);
         }
