+++ core/src/main/java/net/techcable/sonarpet/nms/entity/type/EntityMagmaCubePet.java
@@ -10,15 +10,7 @@ import net.techcable.sonarpet.nms.NMSInsentientEntity;
 
 @EntityPetType(petType = PetType.MAGMACUBE)
 public class EntityMagmaCubePet extends AbstractEntitySlimePet implements IEntityMagmaCubePet {
-    private final NMSInsentientEntity entity;
-
     protected EntityMagmaCubePet(IPet pet, NMSInsentientEntity entity) {
-        super(pet);
-        this.entity = entity;
-    }
-
-    @Override
-    public NMSInsentientEntity getEntity() {
-        return entity;
+        super(pet, entity);
     }
 }
