+++ core/src/main/java/net/techcable/sonarpet/nms/entity/type/EntityPigZombiePet.java
@@ -4,12 +4,9 @@ import com.dsh105.echopet.compat.api.entity.EntityPetType;
 import com.dsh105.echopet.compat.api.entity.IPet;
 import com.dsh105.echopet.compat.api.entity.PetType;
 import com.dsh105.echopet.compat.api.entity.type.nms.IEntityPigZombiePet;
-import net.techcable.sonarpet.nms.INMS;
 
 import net.techcable.sonarpet.nms.entity.AbstractEntityZombiePet;
 import net.techcable.sonarpet.nms.NMSInsentientEntity;
-import net.techcable.sonarpet.nms.switching.ZombieType;
-import net.techcable.sonarpet.nms.switching.EntitySwitchReason;
 import net.techcable.sonarpet.utils.NmsVersion;
 import net.techcable.sonarpet.utils.Versioning;
 
@@ -24,11 +21,7 @@ public class EntityPigZombiePet extends AbstractEntityZombiePet implements IEnti
     @SuppressWarnings("deprecation") // We only do it when it's safe
     public void setVillager(boolean flag) {
         if (Versioning.NMS_VERSION.compareTo(NmsVersion.v1_11_R1) >= 0) {
-            entity = INMS.getInstance().switchType(
-                    entity,
-                    EntitySwitchReason.ZOMBIE_SWITCH,
-                    ZombieType.VILLAGER
-            );
+            throw new UnsupportedOperationException("TODO");
         } else {
             getBukkitEntity().setVillager(flag);
         }
