+++ core/src/main/java/net/techcable/sonarpet/nms/entity/type/EntityZombiePet.java
@@ -4,14 +4,12 @@ import com.dsh105.echopet.compat.api.entity.EntityPetType;
 import com.dsh105.echopet.compat.api.entity.IPet;
 import com.dsh105.echopet.compat.api.entity.PetType;
 import com.dsh105.echopet.compat.api.entity.type.nms.IEntityZombiePet;
-import net.techcable.sonarpet.nms.INMS;
 
 import net.techcable.sonarpet.nms.entity.AbstractEntityZombiePet;
 import net.techcable.sonarpet.nms.NMSInsentientEntity;
 import net.techcable.sonarpet.nms.entity.generators.EntityUndeadPetGenerator;
 import net.techcable.sonarpet.nms.entity.generators.GeneratorClass;
-import net.techcable.sonarpet.nms.switching.ZombieType;
-import net.techcable.sonarpet.nms.switching.EntitySwitchReason;
+import com.dsh105.echopet.compat.api.entity.ZombieType;
 import net.techcable.sonarpet.utils.NmsVersion;
 import net.techcable.sonarpet.utils.Versioning;
 
@@ -28,11 +26,7 @@ public class EntityZombiePet extends AbstractEntityZombiePet implements IEntityZ
     @Override
     public void setZombieType(ZombieType type) {
         if (Versioning.NMS_VERSION.compareTo(NmsVersion.v1_11_R1) >= 0) {
-            entity = INMS.getInstance().switchType(
-                    entity,
-                    EntitySwitchReason.ZOMBIE_SWITCH,
-                    type
-            );
+            throw new UnsupportedOperationException("TODO");
         } else {
             switch (type) {
                 case REGULAR:
